refactor(LAYOUT): HEADER & FOOTER DIR 구조 추가
