+++ src/components/layout/footer/footer.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import Paper from '@material-ui/core/Paper';
 import { makeStyles } from '@material-ui/core/styles';
 
-import Contact from '../contact';
+import Contact from '../../contact';
 
 interface FooterProps {
 }
