+++ src/components/layout/layout.tsx
@@ -2,8 +2,8 @@ import React from 'react';
 import { Container } from '@material-ui/core';
 import { makeStyles } from '@material-ui/core/styles';
 
-import Header from './header';
-import Footer from './footer';
+import Header from './header/header';
+import Footer from './footer/footer';
 
 interface LayoutProps {
     children : React.ReactChild
